requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
beautifulsoup4==4.12.2
python-dotenv==1.0.0
rapidfuzz==3.5.2
whoosh==2.7.4
psutil==5.9.6
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.1
